Add desktop settings page
